scientific notation/int fix

+++ R/StyleClass.R
@@ -68,8 +68,8 @@ Style$methods(show = function(print = TRUE){
   
   validNumFmt <- c("GENERAL", "NUMBER", "CURRENCY", "ACCOUNTING", "DATE", "TIME", "PERCENTAGE", "SCIENTIFIC", "TEXT")
   
-  if(as.numeric(numFmt$numFmtId) %in% unlist(numFmtMapping)){
-    numFmtStr <- validNumFmt[unlist(numFmtMapping) == as.numeric(numFmt$numFmtId)]
+  if(as.integer(numFmt$numFmtId) %in% unlist(numFmtMapping)){
+    numFmtStr <- validNumFmt[unlist(numFmtMapping) == as.integer(numFmt$numFmtId)]
   }else{
     numFmtStr <- sprintf('"%s"', numFmt$formatCode)
   }
